010116#05: fix a problem in temp file creating & combine syntax with core

+++ core/general.cpp
@@ -32,7 +32,8 @@ void cleanAll()
 	}
 
 	// remove(INK_TMP_PATH);
-	removeDir(INK_TMP_PATH);
+	if (isDirExist(INK_TMP_PATH))
+		removeDir(INK_TMP_PATH);
 }
 
 Ink_Argument::~Ink_Argument()
